third_party/WebKit/LayoutTests/imported/wpt/html/semantics/forms/form-submission-0/getactionurl-expected.txt

This is a testharness.js-based test.
FAIL Navigating to URL with a data scheme assert_equals: expected (string) "hello world" but got (object) null
Harness: the test ran to completion.
